Added support for status line and RSSI Icon
Added support to utilize the 4 th line as a status line instead of the
Connection line..  This might be usefull to break out these changes into
their own class.

Added an Icon for the RSSI.  This comprises of four bars and an antenna
symbol.  This will be place all the way to the left on the first row.
You can also disply the RSSI value, It will display after the Icon if it
is displayed.